// typings/Diamond.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */
import {
  BaseContract,
  BigNumber,
  BytesLike,
  CallOverrides,
  PopulatedTransaction,
  Signer,
  utils,
} from "ethers";
import { FunctionFragment, Result } from "@ethersproject/abi";
import { Listener, Provider } from "@ethersproject/providers";
import { TypedEventFilter, TypedEvent, TypedListener, OnEvent } from "./common";

export interface DiamondInterface extends utils.Interface {
  functions: {
    "c_e3a4c27e(bytes8)": FunctionFragment;
    "c_falsee3a4c27e(bytes8)": FunctionFragment;
    "c_truee3a4c27e(bytes8)": FunctionFragment;
  };

  encodeFunctionData(
    functionFragment: "c_e3a4c27e",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "c_falsee3a4c27e",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "c_truee3a4c27e",
    values: [BytesLike]
  ): string;

  decodeFunctionResult(functionFragment: "c_e3a4c27e", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "c_falsee3a4c27e",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "c_truee3a4c27e",
    data: BytesLike
  ): Result;

  events: {};
}

export interface Diamond extends BaseContract {
  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  interface: DiamondInterface;

  queryFilter<TEvent extends TypedEvent>(
    event: TypedEventFilter<TEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TEvent>>;

  listeners<TEvent extends TypedEvent>(
    eventFilter?: TypedEventFilter<TEvent>
  ): Array<TypedListener<TEvent>>;
  listeners(eventName?: string): Array<Listener>;
  removeAllListeners<TEvent extends TypedEvent>(
    eventFilter: TypedEventFilter<TEvent>
  ): this;
  removeAllListeners(eventName?: string): this;
  off: OnEvent<this>;
  on: OnEvent<this>;
  once: OnEvent<this>;
  removeListener: OnEvent<this>;

  functions: {
    c_e3a4c27e(
      c__e3a4c27e: BytesLike,
      overrides?: CallOverrides
    ): Promise<[void]>;

    c_falsee3a4c27e(
      c__e3a4c27e: BytesLike,
      overrides?: CallOverrides
    ): Promise<[boolean]>;

    c_truee3a4c27e(
      c__e3a4c27e: BytesLike,
      overrides?: CallOverrides
    ): Promise<[boolean]>;
  };

  c_e3a4c27e(c__e3a4c27e: BytesLike, overrides?: CallOverrides): Promise<void>;

  c_falsee3a4c27e(
    c__e3a4c27e: BytesLike,
    overrides?: CallOverrides
  ): Promise<boolean>;

  c_truee3a4c27e(
    c__e3a4c27e: BytesLike,
    overrides?: CallOverrides
  ): Promise<boolean>;

  callStatic: {
    c_e3a4c27e(
      c__e3a4c27e: BytesLike,
      overrides?: CallOverrides
    ): Promise<void>;

    c_falsee3a4c27e(
      c__e3a4c27e: BytesLike,
      overrides?: CallOverrides
    ): Promise<boolean>;

    c_truee3a4c27e(
      c__e3a4c27e: BytesLike,
      overrides?: CallOverrides
    ): Promise<boolean>;
  };

  filters: {};

  estimateGas: {
    c_e3a4c27e(
      c__e3a4c27e: BytesLike,
      overrides?: CallOverrides
    ): Promise<BigNumber>;

    c_falsee3a4c27e(
      c__e3a4c27e: BytesLike,
      overrides?: CallOverrides
    ): Promise<BigNumber>;

    c_truee3a4c27e(
      c__e3a4c27e: BytesLike,
      overrides?: CallOverrides
    ): Promise<BigNumber>;
  };

  populateTransaction: {
    c_e3a4c27e(
      c__e3a4c27e: BytesLike,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;

    c_falsee3a4c27e(
      c__e3a4c27e: BytesLike,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;

    c_truee3a4c27e(
      c__e3a4c27e: BytesLike,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;
  };
}